fix: push

+++ config/filesystems.php
@@ -63,7 +63,7 @@ return [
             'password' => env('FTP_PASSWORD'),
 
             // Optional
-            'port' => env('FTP_PORT', 21),
+            'port' => (int)env('FTP_PORT', 21),
             'root'     => env('FTP_ROOT', '/web'),
             'passive'  => true,
             'ssl'      => false,
